Version 4.0
Fixes:
Fix the problems in using the taglist plugin with Vim7 tabs. When
Tlist_File_Fold_Auto_Close is set, close the tag fold when leaving a
buffer.  When jumping to a selected tag, if the file is opened in more
than one window, use the previous window from which the user entered
the taglist window.
New features:
Support for displaying the tag prototype as a tooltip.  Support for
specifying additional commands after the taglist commands using the
bar separator.